requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
flask==3.0.0
python-dotenv==1.0.0
fastmcp==2.11.3